+++ Magic/CHANGELOG.md
@@ -13,6 +13,7 @@
 # 8.4.1
 
  - Fix reloading configs setting SP earn cooldown to zero for all online players
+ - Update/Fix Towny integration
 
 # 8.4